composite/appointment/requirements.txt
fastapi==0.134.0
pydantic
uvicorn
httpx
aio-pika
python-dotenv
